// src/generated/api/review/models/reviewStats.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * ReviewService
 * OpenAPI spec version: 1.0.0
 */

export interface ReviewStats {
	/**
	 * @minimum 1
	 * @maximum 5
	 */
	docencia?: number;
	/**
	 * @minimum 1
	 * @maximum 5
	 */
	vibes?: number;
	/**
	 * @minimum 1
	 * @maximum 5
	 */
	relevancia?: number;
	/**
	 * @minimum 1
	 * @maximum 5
	 */
	carga?: number;
	/**
	 * @minimum 1
	 * @maximum 5
	 */
	dificultad?: number;
}